2007-03-13	Ralf Corsépius <ralf.corsepius@rtems.org>
	* score/include/rtems/score/chain.h, score/src/chain.c:
	Change _Chain_Initialize to using size_t.

+++ cpukit/score/include/rtems/score/chain.h
@@ -109,8 +109,8 @@ typedef struct {
 void _Chain_Initialize(
   Chain_Control *the_chain,
   void          *starting_address,
-  uint32_t       number_nodes,
-  uint32_t       node_size
+  size_t         number_nodes,
+  size_t         node_size
 );
 
 /**
